Move rate limit key tests to package level
This reflects their location in the code.
