+++ tests/clients/redis/test_redis_client.py
@@ -5,10 +5,6 @@ from unittest.mock import Mock, call
 import pytest
 from freezegun import freeze_time
 
-from notifications_utils.clients.redis import (
-    daily_limit_cache_key,
-    rate_limit_cache_key,
-)
 from notifications_utils.clients.redis.redis_client import (
     RedisClient,
     prepare_value,
@@ -146,15 +142,6 @@ def test_should_call_set_if_enabled(mocked_redis_client):
 def test_should_call_get_if_enabled(mocked_redis_client):
     assert mocked_redis_client.get('key') == 100
     mocked_redis_client.redis_store.get.assert_called_with('key')
-
-
-def test_daily_limit_cache_key(sample_service):
-    with freeze_time("2016-01-01 12:00:00.000000"):
-        assert daily_limit_cache_key(sample_service.id) == '{}-2016-01-01-count'.format(sample_service.id)
-
-
-def test_rate_limit_cache_key(sample_service):
-    assert rate_limit_cache_key(sample_service.id, 'TEST') == '{}-TEST'.format(sample_service.id)
 
 
 @freeze_time("2001-01-01 12:00:00.000000")
